refactor:organização de dependências
