+++ main.py
@@ -1,9 +1,9 @@
-from nokia import *
+# from nokia_ssh import *
 
-def main():
-    # Estabelece conexão
-    connection = login_olt_ssh()
-    list_unauthorized(connection)
+# def main():
+#     # Estabelece conexão
+#     connection = login_olt_ssh()
+#     list_unauthorized(connection)
 
-if __name__ == "__main__":
-    main()
+# if __name__ == "__main__":
+#     main()
