+++ nokia/nokia_tl1.py
@@ -5,7 +5,6 @@ import logging
 # from collections import defaultdict
 import time
 
-
 def setup_logging():
     """Configura o sistema de logging com rotação de arquivos"""
     logging.basicConfig(
@@ -17,6 +16,7 @@ def setup_logging():
     logging.info("="*50)
     logging.info("Iniciando nova sessão de conexão OLT")
     logging.info("="*50)
+
 
 def log_interaction(action, details="", level="info"):
     """Função padronizada para registro de logs"""
@@ -33,74 +33,6 @@ def log_interaction(action, details="", level="info"):
 # Carrega variáveis do arquivo .env
 load_dotenv()
 setup_logging()
-
-def login_olt_ssh():
-    """Estabelece conexão SSH com a OLT"""
-    try:
-        log_interaction("Iniciando conexão SSH", "Obtendo variáveis de ambiente")
-        
-        # Obtém variáveis de ambiente
-        ssh_user = os.getenv('SSH_USER')
-        olt_ip = os.getenv('LAB_IP')
-        ssh_password = os.getenv('SSH_PASSWORD')
-        port = os.getenv('PORT')
-        
-        # Validação das variáveis
-        if not all([ssh_user, olt_ip, ssh_password]):
-            error_msg = "Variáveis de ambiente não configuradas corretamente"
-            log_interaction("Erro de configuração", error_msg, "error")
-            raise ValueError(error_msg)
-       
-        log_interaction("Conectando à OLT", f"Usuário: {ssh_user} | OLT: {olt_ip}")
-        # print(f"Usuário: {ssh_user}, IP: {olt_ip}, Porta: {port}")
-        # Conexão SSH
-        child = pexpect.spawn(f"ssh {ssh_user}@{olt_ip} -p {port}", encoding='utf-8', timeout=30)
-
-        log_interaction("SSH iniciado", "Aguardando prompt de senha", "debug")
-
-        index = child.expect(["password:", "Are you sure you want to continue connecting", pexpect.TIMEOUT], timeout=10)
-        
-        if index == 1:
-            log_interaction("SSH", "Primeira conexão - aceitando certificado", "debug")
-            child.sendline("yes")
-            child.expect("password:")
-        
-        log_interaction("SSH", "Enviando credenciais de acesso", "debug")
-        child.sendline(ssh_password)
-        login_success = child.expect([r"typ:isadmin>#", pexpect.TIMEOUT, pexpect.EOF], timeout=10)
-
-        if login_success == 0:
-            log_interaction("Conexão estabelecida", f"Conectado à OLT {olt_ip}")
-            print(f"✅ Conectado com sucesso à OLT {olt_ip}")
-
-        else:
-            log_interaction("Erro SSH", "Não foi possível autenticar na OLT", "error")
-            print("❌ Falha na autenticação")
-            return None
-        child.sendline("environment inhibit-alarms")
-        child.expect("#")
-        child.sendline("exit")
-        child.expect("#")
-        log_interaction("Alarmes desativados.")
-        return child
-    
-
-    except pexpect.EOF:
-        log_interaction("Erro SSH", "Conexão fechada inesperadamente", "error")
-        print("❌ Erro: Conexão foi fechada antes do login.")
-        
-    except pexpect.exceptions.ExceptionPexpect as e:
-        error_msg = f"Falha na conexão SSH: {str(e)}"
-        log_interaction("Erro SSH", error_msg, "error")
-        print(error_msg)
-        return None
-        
-    except Exception as e:
-        error_msg = f"Erro inesperado: {str(e)}"
-        log_interaction("Erro geral", error_msg, "error")
-        print(error_msg)
-        return None
-
 
 def login_olt_tl1(): 
     """Estabelece conexão TL1 com a OLT"""
@@ -190,21 +122,4 @@ def login_olt_tl1():
         log_interaction("Erro geral TL1", error_msg, "error")
         print(error_msg)
         return None
-    
-def list_unauthorized(child):
-            child.sendline('show pon unprovision-onu')
-            time.sleep(3)
-            child.sendline('logout')
-            child.terminate()
-            parse = str(child.readlines()[7::])
-            parse_split = parse.split(",")
-            
-            for count,line in enumerate(parse_split):
-                if "1/1/" in line:
-                    pos_slot = line.find('1/1/')
-                    dados_temp = line[pos_slot:].split(" ")
-                    dados = list(filter(None,dados_temp))
-                    print(f"Slot: {dados[0].split('/')[2]} Pon: {dados[0].split('/')[3]} Serial: {dados[1]}")
-                    continue
-                continue
